+++ doc/WHATSNEW.md
@@ -1,6 +1,8 @@
 Version ?? (unreleased)
 -----------------------
 
+- New: new dbconsole command: `user:drop`
+
 - Improved: results of SELECT commands are now displayed using table formatting for MariaDB and MySQL databases
   when running dbconsole command `sql:execute`